Default data type confusion
Since there is no way of knowing beforehand if the generated Petri nets
are going to be bounded, and even if they are bounded at low values, it
makes little sense to save bits at this starting point. I could also
just keep the previous value for ints and then increase their size when
solving the nets, so this could be reveresed later.

+++ spn_generate_random.m
@@ -10,7 +10,7 @@
 %%   cm: The compound matrix (inflows (pn x tn); outflows (pn x tn); M_0 (tn x 1))
 %%   lambda: A vector with the lambda for each transition.
 function [cm, lambda] = spn_generate_random(pn, tn, prob, max_lambda)
-  cm = zeros(pn, 2*tn + 1, "uint8");
+  cm = zeros(pn, 2*tn + 1, "uint32");
 
   % Nodes are enumerated sequentially, with places preceding transitions. That's
   % why we add the number of places afterwards.
